Enable safe optimizations in debug builds.
Disallowing all optimizations often make things unbearably slow, and
there is an option to allow all `safe` optimizations. Safe here means
optimizations that will not interfere with debugging.

Small note: This is not guaranteed to be available for _all_ compilers.
It's definitely there for GCC and most common alternatives, but not
guaranteed for all. But this is the case for all these options that are
set, so making this more general will (IMO) cause more bloat
than the good it will provide.